backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.1
pydantic-settings==2.1.0
uvicorn==0.25.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
alembic==1.13.1
pytest==7.4.3
httpx==0.27.0
email-validator==2.1.0
qrcode==7.4.2
pillow==10.2.0 